+++ lib/src/schemas/schemas.dart
@@ -1,8 +1,10 @@
 export 'schema.dart';
+export 'async_refined_schema.dart';
 export 'z_bool.dart';
 export 'z_coerce_container.dart';
 export 'z_date.dart';
 export 'z_default.dart';
+export 'z_discriminated_union.dart';
 export 'z_double.dart';
 export 'z_enum.dart';
 export 'z_file.dart';
@@ -14,5 +16,6 @@ export 'z_map.dart';
 export 'z_nullable.dart';
 export 'z_num.dart';
 export 'z_optional.dart';
+export 'piped_schema.dart';
 export 'z_string.dart';
 export 'z_union.dart';